latest revision from edits during PM

+++ sortio.h
@@ -27,18 +27,18 @@ class sortio_Class {
 
 
  private:
-  bool master;			// master task?
-  bool initialized;		// class initialized?
-  bool override_numfiles;       // Override num_files setting?
-  bool random_read_offset;      // Randomly change rank ordering for read to minimize cache effects?
-  int  nio_tasks;		// number of dedicated MPI I/O tasks
-  int  io_rank;			// MPI rank of local I/O task
-  int  num_files_total;		// total # of input files to sort
-  int  num_records_read;	// total # of records read locally
-  MPI_Comm IO_COMM;		// MPI communicator for raw I/O tasks
-  std::string basename;		// input file basename
-  std::string indir;		// input directory
-  GRVY::GRVY_Timer_Class gt;    // performance timer
+  bool master;			   // master task?
+  bool initialized;		   // class initialized?
+  bool override_numfiles;          // Override num_files setting?
+  bool random_read_offset;         // Randomly change rank ordering for read to minimize cache effects?
+  int  nio_tasks;		   // number of dedicated MPI I/O tasks
+  int  io_rank;			   // MPI rank of local I/O task
+  int  num_files_total;		   // total # of input files to sort
+  unsigned long num_records_read;  // total # of records read locally
+  MPI_Comm IO_COMM;		   // MPI communicator for raw I/O tasks
+  std::string basename;		   // input file basename
+  std::string indir;		   // input directory
+  GRVY::GRVY_Timer_Class gt;       // performance timer
 
   unsigned char rec_buf[REC_SIZE];
